requirements.txt
matplotlib~=3.5.1
numpy~=1.21.5
mysql-connector-python~=8.0.18
future~=0.18.2
tkcalendar~=1.6.1